% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_neat.R
\name{m_neat}
\alias{m_neat}
\title{Neat M and SD}
\usage{
m_neat(values, round_to = 0, new_name = NULL, group_by = NULL,
  medians = F)
}
\description{
This function gives means (or medians) and SDs per group for given variable. Mainly for use in the table_neat() function.
}
\examples{
m_neat()
}
\keyword{table}
