% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mon_neat.R
\name{mon_deg2pix}
\alias{mon_deg2pix}
\title{Degrees to pixels}
\usage{
mon_deg2pix(mon_obj, degrees)
}
\arguments{
\item{mon_obj}{Monitor object, as assigned with \code{\link{mon_neat}}.}

\item{pixels}{Number of degrees.}
}
\value{
Number of pixels for the given number of degrees.
}
\description{
Given a specific monitor object, converts visual angle degrees
  to pixels.
}
\examples{
# assign monitor with 57 cm distance, and screen width 52 cm and 1920 pixels
my_mon = mon_neat(distance = 57, mon_width_cm = 52, mon_width_pixel = 1920)

# convert 0.8 degrees of visual angle to pixels, for the specified monitor
mon_deg2pix( my_mon, 0.8 ) # returns 29.38645 (pixels)
}
\seealso{
\code{\link{mon_neat}}, \code{\link{mon_deg2pix}},
  \code{\link{mon_params}}
}
