% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mon_neat.R
\name{mon_params}
\alias{mon_params}
\title{Monitor parameters}
\usage{
mon_params(mon_obj)
}
\arguments{
\item{mon_obj}{Monitor object, as assigned with \code{\link{mon_neat}}.}
}
\value{
Returns nothing, just prints all parameters.
}
\description{
Prints the parameters of a given monitor object.
}
\examples{
# assign monitor with 57 cm distance, and screen width 52 cm and 1920 pixels
my_mon = mon_neat(distance = 57, mon_width_cm = 52, mon_width_pixel = 1920)

# print the above given parameters
mon_params( my_mon )
}
\seealso{
\code{\link{mon_neat}}, \code{\link{mon_conv}}
}
