% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_path.R
\name{script_path}
\alias{script_path}
\title{Script Path}
\usage{
script_path(subdir = "")
}
\arguments{
\item{subdir}{String, optional. Subdirectory relative to the script's path.}
}
\value{
Script file's path as string. If \code{subdir} is given, it is
  appended to the original path.
}
\description{
Gives, in RStudio, the path to the script file in which it is
  executed.
}
\examples{
\donttest{
setwd( script_path() ) # sets working directory to the script's path, e.g. "C:/script_folder/"
}

script_path('my_subdir/misc/') # returns "C:/script_folder/my_subdir/misc/"
}
