% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_ci.R
\name{sd_ci}
\alias{sd_ci}
\title{Confidence Interval of Standard Deviation}
\usage{
sd_ci(x, ci = 0.95)
}
\arguments{
\item{x}{Numeric vector.}

\item{ci}{Numeric; confidence level for returned CI.}
}
\value{
SD confidence interval (as named vector).
}
\description{
Calculates the SD confidence interval of a vector of numbers.
}
\examples{
myvec = c(11, 15, 19, 43, 53, -4, 34, 8, 33, -1, 54 )
sd_ci( myvec )
sd_ci( myvec, ci = .80 )

}
\seealso{
\code{\link{mean_ci}}, \code{\link{plot_neat}}
}
