% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error of Mean}
\usage{
se(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Standard error.
}
\description{
Simply calculates the standard error of a vector of numbers.
}
\examples{
se( c(11, 15, 19, 43, 53, -4, 34, 8, 33, -1, 54 ) )

}
\seealso{
\code{\link{mean_ci}}, \code{\link{plot_neat}}
}
