% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatStats.R
\name{anova_neat}
\alias{anova_neat}
\title{Neat ANOVA}
\usage{
anova_neat(data_long, value_col, id_col, between_vars = NULL,
  within_vars = NULL)
}
\description{
This function gives thorough ANOVA results including CIs and BFs.
}
\examples{
anova_neat()
}
\keyword{anova}
