% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_neat.R
\name{corr_neat}
\alias{corr_neat}
\title{Neat correlation}
\usage{
corr_neat(var1, var2, ci = 0.95, bf_added = T, direction = "",
  round_r = 3, for_table = F)
}
\description{
This function gives Pearson correlation results including CIs and BFs.
}
\examples{
corr_neat()
}
\keyword{correlation}
