% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dems_neat.R
\name{dems_neat}
\alias{dems_neat}
\title{Demographics}
\usage{
dems_neat(data_per_subject, group_by = NULL, percent = T,
  round_perc = 0)
}
\description{
Prints participant count, age, gender, from given dataset.
}
\examples{
dems_neat()
}
