% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excl_neat.R
\name{excl_neat}
\alias{excl_neat}
\title{Exclusion}
\usage{
excl_neat(dat, filt, group_by = NULL, sort_by = "exclusion", hush = FALSE)
}
\arguments{
\item{dat}{Data frame to be filtered.}

\item{filt}{An expression to filter, by column values, the \code{dat}
data frame. The expression should use column names alone; see Examples.}

\item{group_by}{String, or vector of strings: the name(s) of the column(s) in
the \code{dat} data frame, containing the vector(s) of factors by which the
printed counts are grouped.}

\item{sort_by}{String; specifies whether the printed counts should be sorted
by exclusion (default; \code{"exclusion"} or its short forms, e.g.
\code{"excl"}), or by the factors given for \code{group_by} (for this, give
any other string, e.g. \code{"conditions"}). If \code{NULL} (default).}

\item{hush}{Logical. If \code{TRUE}, prevents printing counts to console.}
}
\value{
A data frame with the statistics per group, with a single column
  (\code{"aggr_group"}) indicating the grouping.
}
\description{
Filters dataset by rows (normally: subjects, observations) and
  prints the numbers of excluded rows and remaining rows.
}
\examples{

}
\seealso{
\code{\link{table_neat}} to create full tables using multiple
  variables
}
