% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_neat.R
\name{m_neat}
\alias{m_neat}
\title{Neat Descriptives}
\usage{
m_neat(values, round_to = 0, new_name = NULL, group_by = NULL,
  medians = F)
}
\arguments{
\item{values}{A vector of numbers from which the statistics are to be calculated.}

\item{round_to}{Number of digits after the decimal point to round to.}

\item{new_name}{String. A new name for the variable to be used as column title.}

\item{group_by}{A vector of factors by which the statistics are grouped.}

\item{medians}{Logical. If \code{TRUE}, medians are calculated, otherwise means.}
}
\value{
A data frame with the statistics per group. Furthermore,
  prints statistics per group, unless used within the
  \code{\link{table_neat}} function, in which case nothing is printed.
}
\description{
Returns means (or medians) and SDs per group for given variable.
    Primarily for use in the \code{\link{table_neat}} function.
}
\examples{
data("mtcars") # load base R example dataset

res1 = m_neat(mtcars$wt ) # overall means and SDs for wt (Weight)

res2 = m_neat(mtcars$wt, 2, new_name = 'weight') # rounded to 2 and renamed

res3 = m_neat(mtcars$wt, 2, group_by = mtcars$cyl) # grouped by cyl (Number of cylinders)

res4 = m_neat(mtcars$wt, 2, group_by = mtcars$cyl, medians = T) # medians
}
\seealso{
\code{\link{table_neat}} to create full tables using multiple variables
}
