% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Confidence Interval of Mean}
\usage{
mean_ci(x, ci = 0.95)
}
\arguments{
\item{x}{Numeric vector.}

\item{ci}{Numeric; confidence level for returned CI.}
}
\value{
Confidence interval (as named vector).
}
\description{
Calculates confidence interval of a vector of numbers.
}
\examples{
mean_ci( c(11, 15, 19, 43, 53, -4, 34, 8, 33, -1, 54 ) )
mean_ci( c(11, 15, 19, 43, 53, -4, 34, 8, 33, -1, 54 ), ci = .80 )

}
\seealso{
\code{\link{se}}
}
