% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatStats.R
\docType{package}
\name{neatStats-package}
\alias{neatStats}
\alias{neatStats-package}
\title{neatStats: an R Package for Neat and Painless Statistical Reporting}
\description{
This package focuses on user-friendly, simple functions to produce clear but comprehensive statistics, primarily for publication in psychological science.
Most functions are wrappers for other packages, but there are various additions as well.
}
\author{
\strong{Maintainer}: Gáspár Lukács \email{lkcsgaspar@gmail.com} (0000-0001-9401-4830)

}
\keyword{internal}
