% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_neat.R
\name{path_neat}
\alias{path_neat}
\title{Script Path}
\usage{
path_neat(subdir = "")
}
\arguments{
\item{subdir}{String, optional. Subdirectory relative to the script's path.}
}
\value{
Script file's path as string. If \code{subdir} is given, it is
  appended to the original path.
}
\description{
Gives, in RStudio, the path to the script file in which it is
  executed.
}
\examples{

# assuming the given script is at path "C:/script_folder/"
path_neat('') # returns "C:/script_folder/"
path_neat('my_subdir/misc/') # returns "C:/script_folder/my_subdir/misc/"

# Note: the returned string can be used as argument for base::setwd()
# e.g. setwd( path_neat() ) # sets working directory to the script's path

}
