% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neat.R
\name{plot_neat}
\alias{plot_neat}
\title{Plots of Means}
\usage{
plot_neat(data_per_subject, values, between_vars = NULL,
  within_ids = NULL, factor_names = NULL, value_names = NULL,
  y_title = NULL, reverse = FALSE, panels = NULL, type = "line",
  dodge = NULL, bar_colors = c("#555555", "#AAAAAA"), method = mean,
  eb_method = stats::sd)
}
\description{
Plots for factorial designs.
}
