% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props_neat.R
\name{props_neat}
\alias{props_neat}
\title{Neat comparison of two proportions}
\usage{
props_neat(case1, case2, n1, n2, greater = "", ci = NULL,
  bf_added = T, h_added = F, for_table = F)
}
\description{
This function compares two independent proportions; including CIs and BFs.
}
\examples{
props_neat()
}
\keyword{proportions}
