% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_neat.R
\name{rbind_loop}
\alias{rbind_loop}
\title{Merge by Columns in Loops}
\usage{
rbind_loop(merged, to_merge)
}
\arguments{
\item{merged}{The name of the data frame for the merged data (without
quotes).}

\item{to_merge}{Data frame to be merged. (Normally a single-row data frame.)}
}
\description{
Merges raws by columns in a loop using the \code{\link{enum}}
  function. On first iteration, indicated by \code{\link{enum}}, initiates a
  new data frame with the data to merge as first row. On all following
  iterations, adds data to merge as subsequent rows (using
  \code{\link[plyr:rbind.fill]{plyr:rbind.fill}}).
}
\details{
See an extensive example via https://github.com/gasparl/neatstats.
}
\examples{

my_vector = c('aa', 'bb', 'cxyz', 'last')
for (elem in enum(my_vector)) {
    cat(elem, fill = TRUE)
    datfram = data.frame(item = elem[2],
                         number = elem[1],
                         whatever = paste0('number (', elem[1], ')'))
    rbind_loop(merged_data, datfram)
}
# merged_data now contained all merged rows
print(merged_data)

# an extensive example to show how to collect and aggregate raw data is
# available via the README file at the repository:
# https://github.com/gasparl/neatstats

}
\seealso{
\code{\link{enum}}
}
