% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro.R
\name{ro}
\alias{ro}
\title{Neat rounding}
\usage{
ro(num, round_to = 2, leading_zero = TRUE)
}
\arguments{
\item{num}{Number to be rounded.}

\item{round_to}{Number of fractional digits (i.e., digits after the decimal
point), to round to.}

\item{leading_zero}{Logical, \code{TRUE} by default. If \code{FALSE}, omits
leadig zero (e.g. returns ".17" instead of "0.17").}
}
\value{
Number as string: \code{num} rounded to \code{round_to} digits, with
  trailing zeros when applicable.
}
\description{
Rounds a given number to given number of digits after the
  decimal point, returning it as string, with trailing zeros when applicable.
}
\examples{
ro( 1.2345 ) # returns "1.23"

ro( 0.12345, 1 ) # returns "0.1"

ro( 12.3, 4 ) # returns "12.3000"
}
