% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_neat.R
\name{roc_neat}
\alias{roc_neat}
\title{Neat ROC test}
\usage{
roc_neat(roc1, roc2, pair = F, greater = "")
}
\description{
This function gives the result of comparing two ROCs (AUCs).
}
\examples{
roc_neat()
}
\keyword{roc}
