% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatStats.R
\name{script_path}
\alias{script_path}
\title{import data.table plyr pROC MBESS ez psychReport BayesFactor TOSTER ggplot2
Neat path}
\usage{
script_path()
}
\description{
This function gives the path to current script's path in RStudio.
}
\examples{
script_path()
}
\keyword{path}
