% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatStats.R
\name{t_neat}
\alias{t_neat}
\title{Neat t-test}
\usage{
t_neat(var1, var2, pair = F, greater = "", ci = NULL, bf_added = T,
  auc_added = F, r_added = T, for_table = F,
  test_title = "Descriptives:", round_descr = 2, round_auc = 3,
  auc_greater = "")
}
\description{
This function gives thorough t-test results including CIs and BFs.
}
\examples{
t_neat()
}
\keyword{ttest}
