% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatStats.R
\name{t_neat}
\alias{t_neat}
\title{Neat t-test}
\usage{
t_neat(var1, var2, pair = F, greater = "", ci = 0.95,
  for_table = F)
}
\description{
This function gives thorough t-test results including CIs and BFs.
}
\examples{
t_neat()
}
\keyword{ttest}
