% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_neat.R
\name{table_neat}
\alias{table_neat}
\title{Neat Table}
\usage{
table_neat(values_list, group_by = NULL, group_per = "rows",
  to_clipboard = F, medians = NULL)
}
\description{
This function gives a neat M and SD table using m_neat() function(s) as input.
}
\examples{
table_neat()
}
\keyword{table}
